Moves SummaryStatistics since it's only used by the poster generation page
