+++ frontend/src/views/GeneratePosterPage/SummaryStatistics.vue
